+++ source/ci_source/providers/_tests/_bamboo.test.ts
@@ -54,6 +54,7 @@ describe(".isPR", () => {
 
     it("needs to have a PR number", () => {
       let env = Object.assign({}, correctEnv)
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore
       delete env["bamboo_repository_pr_key"]
       const pipelines = new Bamboo(env)